+++ src/Types.hs
@@ -26,7 +26,7 @@ data RepoSource = RepoSource Bool Mirror
   deriving Eq
 
 -- FIXME deal with epel11.x
-data Release = EPEL10Dot Natural | EPEL Natural | EPEL9Next
+data Release = EpelMinor Natural Natural | EPEL Natural | EPEL9Next
              | Centos Natural | Fedora Natural
              | ELN | Rawhide | System
   deriving (Eq, Ord)
@@ -43,7 +43,8 @@ eitherRelease rel =
     "rawhide" -> Right Rawhide
     -- FIXME add proper parsing:
     "epel9-next" -> Right EPEL9Next
-    ('e':'p':'e':'l':'1':'0':'.':n@(_:_)) | all isDigit n -> let br = EPEL10Dot (read n)
+    -- FIXME add epel11.x
+    ('e':'p':'e':'l':'1':'0':'.':n@(_:_)) | all isDigit n -> let br = EpelMinor 10 (read n)
                                                  in Right br
     ('e':'p':'e':'l':n@(_:_)) | all isDigit n -> let br = EPEL (read n)
                                                  in Right br
@@ -72,7 +73,7 @@ eitherRelease rel =
 instance Show Release where
   show Rawhide = "rawhide"
   show (Fedora n) = "f" ++ show n
-  show (EPEL10Dot n) = "epel10." ++ show n
+  show (EpelMinor n m) = "epel" ++ show n ++ "." ++ show m
   show (EPEL n) = (if n <= 6 then "el" else "epel") ++ show n
   show EPEL9Next = "epel9" ++ "-next"
   show ELN = "eln"